+++ apps/frontend/src/pages/overview/overview.tsx
@@ -49,7 +49,7 @@ const Overview: React.FC = () => {
 				{isLoading ? (
 					<Loader color="orange" size="large" />
 				) : (
-					<Courses courses={courses} />
+					<Courses courses={courses} userId={user.id} />
 				)}
 			</div>
 			{isAddCourseModalOpen && <AddCourseModal onClose={handleModalClose} />}